+++ cognite/client/_api/datapoint_tasks.py
@@ -34,7 +34,7 @@ from typing import (
 from google.protobuf.message import Message
 from sortedcontainers import SortedDict, SortedList
 
-from cognite.client.data_classes.datapoints import NUMPY_IS_AVAILABLE, Datapoints, DatapointsArray
+from cognite.client.data_classes.datapoints import NUMPY_IS_AVAILABLE, Aggregate, Datapoints, DatapointsArray
 from cognite.client.utils._auxiliary import import_legacy_protobuf, is_unlimited
 from cognite.client.utils._identifier import Identifier
 from cognite.client.utils._text import convert_all_keys_to_snake_case, to_camel_case, to_snake_case
@@ -115,11 +115,11 @@ class DatapointsPayload(CustomDatapoints):
 class _DatapointsQuery:
     """Internal representation of a user request for datapoints, previously public (before v5)"""
 
-    start: Union[int, str, datetime, None] = None
-    end: Union[int, str, datetime, None] = None
+    start: int | str | datetime | None = None
+    end: int | str | datetime | None = None
     id: Optional[DatapointsId] = None
     external_id: Optional[DatapointsExternalId] = None
-    aggregates: Union[str, List[str], None] = None
+    aggregates: Aggregate | str | list[Aggregate | str] | None = None
     granularity: Optional[str] = None
     limit: Optional[int] = None
     include_outside_points: bool = False